@jeremiahschung Fix date handling in download_count_wheels.py (#532)
* Fix date handling in download_count_wheels.py

The existing script called key.last_modified.utcnow() which always returned the current time regardless of the value in the datetime object last_modified. This resulted in the script downloading all logs every time (the continue statement never executed due to miscalculation in the if statement). Fixed it to pull logs from the previous day only.

* Update download_count_wheels.py

Remove line for debug output